Updated removeFromCartTest
